requirements.txt
python-telegram-bot==21.6
yfinance==0.2.40
apscheduler==3.10.4
requests==2.31.0
pytz==2024.1
